remove duplicated method
