+++ lib/page_magic.rb
@@ -5,11 +5,13 @@ require 'page_magic/wait_methods'
 require 'page_magic/watchers'
 require 'page_magic/session'
 require 'page_magic/session_methods'
-require 'page_magic/instance_methods'
+
 require 'page_magic/elements'
-require 'page_magic/class_methods'
+
 require 'page_magic/element_context'
 require 'page_magic/element'
+require 'page_magic/class_methods'
+require 'page_magic/instance_methods'
 require 'page_magic/drivers'
 
 # module PageMagic - PageMagic is an api for modelling pages in a website.
