+++ spec/page_magic/elements_spec.rb
@@ -162,23 +162,6 @@ module PageMagic
       end
     end
 
-    describe '#element_by_name' do
-      it 'returns the required element' do
-        subject.element :child1, child_selector
-        subject.element :child2, id: 'child 2'
-
-        expect(subject.element_by_name(:child1)).to eq(child_element)
-      end
-
-      context 'element not found' do
-        it 'raises an error' do
-          expected_message = (described_class::ELEMENT_MISSING_MSG % :child)
-          command = proc { subject.element_by_name(:child) }
-          expect(&command).to raise_exception ElementMissingException, expected_message
-        end
-      end
-    end
-
     describe '#element_definitions' do
       it 'should return your a copy of the core definition' do
         subject.text_field :alias, child_selector
