+++ spec/page_magic/instance_methods_spec.rb
@@ -17,6 +17,7 @@ module PageMagic
     it_behaves_like 'session accessor'
     it_behaves_like 'element watcher'
     it_behaves_like 'waiter'
+    it_behaves_like 'element locator'
 
     describe 'execute_on_load' do
       it 'runs the on_load_hook in the context of self' do
